+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## Added
+- "save as" in editor
+- add ability to create activity with empty file
+- add favicon on editor page
+- editor interface customization
+
+## Changed
+- set editor interface language
+
 ## 2.1.0
 ## Added
 - backup api